fixed a compilation warning

+++ corelib/src/VWDictionary.cpp
@@ -202,7 +202,7 @@ void VWDictionary::setFixedDictionary(const std::string & dictionaryPath)
 						{
 							std::getline(file, str);
 							strList = uSplit(str);
-							if(strList.size() == dimension+1)
+							if((int)strList.size() == dimension+1)
 							{
 								//first one is the visual word id
 								std::list<std::string>::iterator iter = strList.begin();
